fix(proxy-identity): use execve call on Linux to start the proxy
